Add the mime database to avoid the icon missing issue at the first boot-up
Althought "update-mime-database /usr/share/mime/" is executed at the first
boot-up, the icons are still missing and a reboot is needed to recover this issue.

We tried to execute "update-mime-database" in the script mk-rootfs-buster.mk but
corresponding files are not generated. Hence we pull those files from the device
and override them in the script to avoid this issue.

Change-Id: Ib42217017f355a9165235cbe286edcedf13fdc8b

Former-commit-id: 33a822aea91079378defc58a53f5d99b5e201a6d